Allow to append specific users to the docker group
Being able to run docker as non-root user can be quite useful on
non-critical systems (e.g. dev or build environments) but it is quite
important to keep in mind that such configuration is quite insecure.
For instance, it will allow to attach with root privileges any files
from the host system into a container.

See also:
- https://www.andreas-jung.com/contents/on-docker-security-docker-group-considered-harmful
- https://docs.docker.com/articles/security/#docker-daemon-attack-surface