Restart rather than reload docker.
state=reloaded doesn't actually have any effect on a running
docker process in ubuntu 14.04. In particular, the /etc/init.d/docker
start script does not execute, so any customised storage options will
not be applied to the running docker process which may result in
a big surprise when the daemon is eventually restarted and images
added to the loopback thin pool device are no longer accessible in
the configured thinpool device.

This change removes the "Reload docker" handler and changes the pre 15.04
handler to use "Restart docker" instead.