Only install linux-image-extra for current kernel
Installing the linux-image-extra virtual package will upgrade
the kernel itself due to its dependencies if it is newer than
the currently installed kernel.

This alone is probably not intended, and it also leads to
problems down the road e. g. with DKMS based installs, because
for the upgraded kernel, there won't be a matching linux-headers
package installed.